driver core: Split devres APIs to device/devres.h
mainline inclusion
from mainline-6.14-rc5
commit a21cad9312767d26b5257ce0662699bb202cdda1
category: feature
bugzilla: https://github.com/RVCK-Project/rvck/issues/243

--------------------------------

device.h is a huge header which is hard to follow and easy to miss
something. Improve that by splitting devres APIs to device/devres.h.

In particular this helps to speedup the build of the code that includes
device.h solely for a devres APIs.

While at it, cast the error pointers to __iomem using IOMEM_ERR_PTR()
and fix sparse warnings.

Acked-by: Arnd Bergmann <arnd@arndb.de>
